ggraph optimisations (first neighbour doesn't point to node itself; cost and distance as float)

+++ mgmap/src/main/java/mg/mgmap/activity/mgmap/features/routing/RoutingProfile.java
@@ -26,10 +26,10 @@ public abstract class RoutingProfile {
         return profileCalculator; // default is no way specific calculator
     }
 
-    public double getCost(WayAttributs wayAttributs, PointModel node, PointModel neighbourNode, boolean primaryDirection){
-        return getCost(wayAttributs, PointModelUtil.distance(node,neighbourNode), PointModelUtil.verticalDistance(node, neighbourNode), primaryDirection);
+    public float getCost(WayAttributs wayAttributs, PointModel node, PointModel neighbourNode, boolean primaryDirection){
+        return getCost(wayAttributs, (float)PointModelUtil.distance(node,neighbourNode), PointModelUtil.verticalDistance(node, neighbourNode), primaryDirection);
     }
-    public double getCost(WayAttributs wayAttributs, double distance, float verticalDistance, boolean primaryDirection){
+    public float getCost(WayAttributs wayAttributs, float distance, float verticalDistance, boolean primaryDirection){
         CostCalculator calculator = this.costCalculator;
         if (wayAttributs != null){
             if (wayAttributs.getDerivedData()==null){
@@ -39,7 +39,7 @@ public abstract class RoutingProfile {
                 calculator = wayCostCalculator;
             }
         }
-        return calculator.calcCosts(distance, verticalDistance, primaryDirection);
+        return (float)calculator.calcCosts(distance, verticalDistance, primaryDirection);
     }
 
     public double heuristic(PointModel node1, PointModel node2){
